Separate list row from picker; fix scroll bug
