+++ packages/core/components/AnnotationFilterForm/index.tsx
@@ -7,7 +7,8 @@ import useAnnotationValues from "./useAnnotationValues";
 import Annotation, { AnnotationName } from "../../entity/Annotation";
 import { AnnotationType } from "../../entity/AnnotationFormatter";
 import FileFilter from "../../entity/FileFilter";
-import ListPicker, { ListItem } from "../ListPicker";
+import ListPicker from "../ListPicker";
+import { ListItem } from "../ListPicker/ListRow";
 import NumberRangePicker from "../NumberRangePicker";
 import SearchBoxForm from "../SearchBoxForm";
 import DateRangePicker from "../DateRangePicker";
